requirements.txt
python-dotenv
google-auth==2.6.0
google-auth-oauthlib==0.5.0
google-auth-httplib2==0.1.0
google-api-python-client==2.39.0
python-telegram-bot==20.0b0
requests==2.27.1